Fix for content shift (issue 59)
I took a look at the problem described in issue #59 and think I found the issue. It appears to be related to the white-space: nowrap value. I have disabled that value for now, and so far the problem seems to be resolved. If needed in the future, it can be re-enabled.